+++ src/main/java/com/member/controller/MemberController.java
@@ -1,20 +1,14 @@
 package com.member.controller;
 
 import com.config.auth.annotation.AuthenticatedMember;
-import com.member.dto.request.LoginRequest;
-import com.member.dto.request.SignUpRequest;
-import com.member.dto.response.LoginResponse;
-import com.member.dto.response.SignUpResponse;
 import com.member.service.MemberService;
 import com.util.cookie.CookieUtils;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -24,25 +18,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class MemberController {
 
     private final MemberService memberService;
-
-    @PostMapping("/signup")
-    public ResponseEntity<SignUpResponse> signUp(@Valid @RequestBody SignUpRequest request) {
-        SignUpResponse signUpResponse = memberService.signUp(request);
-        return ResponseEntity.ok(signUpResponse);
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<LoginResponse> login(@Valid @RequestBody LoginRequest loginRequest,
-                                               HttpServletResponse response) {
-        LoginResponse loginResponse = memberService.login(loginRequest);
-
-        Cookie cookie = CookieUtils.createCookie("token",
-                loginResponse.getAccessToken(),
-                (int) loginResponse.getExpirationTime());
-        response.addCookie(cookie);
-
-        return ResponseEntity.ok(loginResponse);
-    }
 
     @PostMapping("/logout")
     public ResponseEntity<Void> logout(@AuthenticatedMember Long memberId,
